FIX: logs weren't deleting when trash can was pressed

+++ assets/javascripts/discourse/controllers/admin-wizards-api.js.es6
@@ -198,6 +198,9 @@ export default Ember.Controller.extend({
     },
 
     clearLogs() {
+      const name = this.get('api.name');
+      if (!name) return;
+
       ajax(`/admin/wizards/apis/logs/${name.underscore()}`, {
         type: 'DELETE'
       }).catch(popupAjaxError)
